+++ clients/admin-ui/src/features/data-discovery-and-detection/action-center/fields/MonitorTree.tsx
@@ -1,9 +1,9 @@
 import {
-  AntButton as Button,
+  AntBadge as Badge,
   AntFlex as Flex,
   AntTitle as Title,
+  AntTooltip as Tooltip,
   AntTree as Tree,
-  SparkleIcon,
 } from "fidesui";
 import { useRouter } from "next/router";
 import {
@@ -30,11 +30,13 @@ import {
 
 import {
   MAP_DATASTORE_RESOURCE_TYPE_TO_ICON,
+  MAP_TREE_RESOURCE_CHANGE_INDICATOR_TO_STATUS_INFO,
   TREE_NODE_LOAD_MORE_KEY_PREFIX,
   TREE_NODE_LOAD_MORE_TEXT,
   TREE_NODE_SKELETON_KEY_PREFIX,
   TREE_PAGE_SIZE,
 } from "./MonitorFields.const";
+import styles from "./MonitorTree.module.scss";
 import { MonitorTreeDataTitle } from "./MonitorTreeDataTitle";
 import {
   collectAllDescendantUrns,
@@ -42,7 +44,7 @@ import {
   removeChildrenFromNode,
   updateNodeStatus,
 } from "./treeUtils";
-import { CustomTreeDataNode } from "./types";
+import { CustomTreeDataNode, TreeNodeAction } from "./types";
 
 const mapResponseToTreeData = (
   data: Page_DatastoreStagedResourceTreeAPIResponse_,
@@ -54,12 +56,29 @@ const mapResponseToTreeData = (
           treeNode.resource_type as StagedResourceTypeValue
         ]
       : undefined;
+    const statusInfo = treeNode.update_status
+      ? MAP_TREE_RESOURCE_CHANGE_INDICATOR_TO_STATUS_INFO[
+          treeNode.update_status
+        ]
+      : undefined;
+
     return {
       title: treeNode.name,
       key: treeNode.urn,
       selectable: true, // all nodes are selectable since we ignore lowest level descendants in the data
       icon: IconComponent
-        ? () => <IconComponent className="h-full" />
+        ? () => (
+            <Tooltip title={statusInfo?.tooltip}>
+              <Badge
+                className="h-full"
+                offset={[0, 5]}
+                color={statusInfo?.color}
+                dot={!!treeNode.update_status}
+              >
+                <IconComponent className="h-full" />
+              </Badge>
+            </Tooltip>
+          )
         : undefined,
       status: treeNode.update_status,
       isLeaf:
@@ -205,13 +224,12 @@ export interface MonitorTreeRef {
 }
 
 interface MonitorTreeProps {
-  selectedNodeKeys: Key[];
   setSelectedNodeKeys: (keys: Key[]) => void;
-  onClickClassifyButton: () => void;
+  nodeActions: Map<Key, TreeNodeAction>;
 }
 
 const MonitorTree = forwardRef<MonitorTreeRef, MonitorTreeProps>(
-  ({ selectedNodeKeys, setSelectedNodeKeys, onClickClassifyButton }, ref) => {
+  ({ setSelectedNodeKeys, nodeActions }, ref) => {
     const router = useRouter();
     const { errorAlert } = useAlert();
     const monitorId = decodeURIComponent(router.query.monitorId as string);
@@ -528,14 +546,8 @@ const MonitorTree = forwardRef<MonitorTreeRef, MonitorTreeProps>(
       [onLoadData],
     );
 
-    const handleSelect = (
-      _: Key[],
-      info: { selectedNodes: CustomTreeDataNode[] },
-    ) => {
-      const classifyableKeys = info.selectedNodes
-        .filter((node) => node.classifyable)
-        .map((node) => node.key);
-      setSelectedNodeKeys(classifyableKeys);
+    const handleSelect = (keys: Key[]) => {
+      setSelectedNodeKeys(keys);
     };
 
     // Expose the function through the ref
@@ -579,27 +591,30 @@ const MonitorTree = forwardRef<MonitorTreeRef, MonitorTreeProps>(
           showIcon
           showLine
           blockNode
-          rootClassName="h-full overflow-x-hidden"
+          rootClassName={`h-full overflow-x-hidden ${styles["monitor-tree"]}`}
           // eslint-disable-next-line react/no-unstable-nested-components
           titleRender={(node) => (
             <MonitorTreeDataTitle
               node={node}
               treeData={treeData}
               onLoadMore={onLoadMore}
+              actions={nodeActions}
             />
           )}
         />
-        {selectedNodeKeys.length > 0 && (
-          <Flex justify="space-between" align="center">
-            <span>{selectedNodeKeys.length} selected</span>
-            <Button
-              aria-label={`Classify ${selectedNodeKeys.length} Selected Nodes`}
-              icon={<SparkleIcon size={12} />}
-              size="small"
-              onClick={onClickClassifyButton}
-            />
-          </Flex>
-        )}
+        {
+          // Enable when multi select is supported
+          // {selectedNodeKeys.length > 0 && (
+          //   <Flex justify="space-between" align="center">
+          //     <Button
+          //       aria-label={`Classify ${selectedNodeKeys.length} Selected Nodes`}
+          //       icon={<SparkleIcon size={12} />}
+          //       size="small"
+          //       onClick={onClickClassifyButton}
+          //     />
+          //   </Flex>
+          // )}
+        }
       </Flex>
     );
   },